fix: Circular component dependency detected
