+++ src/AeFinder.App.Deploy/IAppResourceLimitProvider.cs
@@ -6,5 +6,5 @@ namespace AeFinder.App.Deploy;
 public interface IAppResourceLimitProvider
 {
     Task<AppResourceLimitDto> GetAppResourceLimitAsync(string appId);
-    Task SetAppPodOperationSnapshotAsync(string appId, string version, AppPodOperationType operationType);
+    
 }
