+++ src/AeFinder.Application.Contracts/Apps/IAppOperationSnapshotProvider.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace AeFinder.Apps;
+
+public interface IAppOperationSnapshotProvider
+{
+    Task SetAppPodOperationSnapshotAsync(string appId, string version, AppPodOperationType operationType);
+}
