+++ src/AeFinder.Application/Apps/AppDeployService.cs
@@ -22,24 +22,24 @@ public class AppDeployService : AeFinderAppService, IAppDeployService
     private readonly IBlockScanAppService _blockScanAppService;
     private readonly IAppDeployManager _appDeployManager;
     private readonly IAppResourceLimitProvider _appResourceLimitProvider;
-    private readonly IKubernetesAppMonitor _kubernetesAppMonitor;
+    private readonly IAppOperationSnapshotProvider _appOperationSnapshotProvider;
 
     public AppDeployService(IClusterClient clusterClient,
         IBlockScanAppService blockScanAppService, IAppDeployManager appDeployManager,
-        IKubernetesAppMonitor kubernetesAppMonitor,IAppResourceLimitProvider appResourceLimitProvider)
+        IAppOperationSnapshotProvider appOperationSnapshotProvider,IAppResourceLimitProvider appResourceLimitProvider)
     {
         _clusterClient = clusterClient;
         _blockScanAppService = blockScanAppService;
         _appDeployManager = appDeployManager;
         _appResourceLimitProvider = appResourceLimitProvider;
-        _kubernetesAppMonitor = kubernetesAppMonitor;
+        _appOperationSnapshotProvider = appOperationSnapshotProvider;
     }
 
     public async Task<string> DeployNewAppAsync(string appId, string version, string imageName)
     {
         var chainIds = await GetDeployChainIdAsync(appId, version);
         var graphqlUrl = await _appDeployManager.CreateNewAppAsync(appId, version, imageName, chainIds);
-        await _appResourceLimitProvider.SetAppPodOperationSnapshotAsync(appId, version, AppPodOperationType.Start);
+        await _appOperationSnapshotProvider.SetAppPodOperationSnapshotAsync(appId, version, AppPodOperationType.Start);
         return graphqlUrl;
     }
 
@@ -47,7 +47,7 @@ public class AppDeployService : AeFinderAppService, IAppDeployService
     {
         var chainIds = await GetSubscriptionChainIdAsync(appId, version);
         await _blockScanAppService.PauseAsync(appId, version);
-        await _appResourceLimitProvider.SetAppPodOperationSnapshotAsync(appId, version, AppPodOperationType.Stop);
+        await _appOperationSnapshotProvider.SetAppPodOperationSnapshotAsync(appId, version, AppPodOperationType.Stop);
         await _appDeployManager.DestroyAppAsync(appId, version, chainIds);
     }
