+++ src/AeFinder.BackgroundWorker.Core/EventHandler/AppUpgradeHandler.cs
@@ -22,11 +22,11 @@ public class AppUpgradeHandler : AppHandlerBase,IDistributedEventHandler<AppUpgr
     private readonly IEntityMappingRepository<AppSubscriptionIndex, string> _appSubscriptionEntityMappingRepository;
     private readonly IEntityMappingRepository<AppSubscriptionPodIndex, string> _appSubscriptionPodEntityMappingRepository;
     private readonly IOrganizationAppService _organizationAppService;
-    private readonly IAppResourceLimitProvider _appResourceLimitProvider;
+    private readonly IAppOperationSnapshotProvider _appOperationSnapshotProvider;
 
     public AppUpgradeHandler(IAppDeployManager kubernetesAppManager,
         IOrganizationAppService organizationAppService,IClusterClient clusterClient,
-        IAppResourceLimitProvider appResourceLimitProvider,
+        IAppOperationSnapshotProvider appOperationSnapshotProvider,
         IEntityMappingRepository<AppInfoIndex, string> appInfoEntityMappingRepository,
         IEntityMappingRepository<AppSubscriptionIndex, string> appSubscriptionEntityMappingRepository,
         IEntityMappingRepository<AppSubscriptionPodIndex, string> appSubscriptionPodEntityMappingRepository)
@@ -34,7 +34,7 @@ public class AppUpgradeHandler : AppHandlerBase,IDistributedEventHandler<AppUpgr
         _clusterClient = clusterClient;
         _organizationAppService = organizationAppService;
         _kubernetesAppManager = kubernetesAppManager;
-        _appResourceLimitProvider = appResourceLimitProvider;
+        _appOperationSnapshotProvider = appOperationSnapshotProvider;
         _appInfoEntityMappingRepository = appInfoEntityMappingRepository;
         _appSubscriptionEntityMappingRepository = appSubscriptionEntityMappingRepository;
         _appSubscriptionPodEntityMappingRepository = appSubscriptionPodEntityMappingRepository;
@@ -49,7 +49,7 @@ public class AppUpgradeHandler : AppHandlerBase,IDistributedEventHandler<AppUpgr
         var historyVersion = eventData.CurrentVersion;
         
         //destory old version pod
-        await _appResourceLimitProvider.SetAppPodOperationSnapshotAsync(appId, eventData.CurrentVersion,
+        await _appOperationSnapshotProvider.SetAppPodOperationSnapshotAsync(appId, eventData.CurrentVersion,
             AppPodOperationType.Stop);
         await _kubernetesAppManager.DestroyAppAsync(appId, eventData.CurrentVersion, eventData.CurrentVersionChainIds);
